Code updates
1. Components: ButtonWithTooltip, SearchBar, DuplicatePageSettings
2. Update added `slug.current` for duplicate document
