+++ studio/components/actions/index.ts
@@ -1,0 +1,3 @@
+export { default as ButtonWithTooltip } from "./ButtonWithTooltip";
+export { default as DuplicatePageSettings } from "./DuplicatePageSettings";
+export { default as SearchBar } from "./SearchBar";
